+++ compiler-plugin/src/main/scala/scalapb/compiler/ProtobufGenerator.scala
@@ -50,13 +50,13 @@ class ProtobufGenerator(
       }
       .add(s"object $name extends ${e.companionExtends.mkString(" with ")} {")
       .indent
-      .seq(e.annotationForTraitList)
+      .seq(e.traitAnnotationList)
       .add(s"sealed trait ${e.recognizedEnum.nameSymbol} extends $name")
       .add(s"implicit def enumCompanion: _root_.scalapb.GeneratedEnumCompanion[$name] = this")
       .print(e.getValues.asScala) { case (p, v) =>
         p.call(generateScalaDoc(v))
           .add("@SerialVersionUID(0L)")
-          .seq(e.annotationForValueList)
+          .seq(e.valueAnnotationList)
           .seq(v.annotationList)
           .add(s"""case object ${v.scalaName.asSymbol} extends ${v.valueExtends
             .mkString(" with ")} {
@@ -67,7 +67,7 @@ class ProtobufGenerator(
                   |""".stripMargin)
       }
       .add("@SerialVersionUID(0L)")
-      .seq(e.annotationListForUnrecognised)
+      .seq(e.unrecognizedAnnotationList)
       .add(s"""${if (e.isPrivateUnrecognized) "private " else ""}final case class Unrecognized(unrecognizedValue: _root_.scala.Int) extends $name(unrecognizedValue) with _root_.scalapb.UnrecognizedEnum
               |
               |lazy val values = scala.collection.immutable.Seq(${e.getValues.asScala