+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
-## [0.11.2] (Unreleased)
+
+## [0.11.4] (Unreleased)
+- Add options for enum annotations (#1155)
+- Add `unrecognized_value_class_is_private` option to make `Unrecognised` class private for safer development. (#1155)
+
+## [0.11.2]
 - Fix bug where generated code would not compile if a message was named Unit (#1126)
 - Add option to block generation/access to field getters (#1098)
-- Add option for enum annotation (#1155)
-- Add `unrecognized_value_class_is_private` option to make `Unrecognised` class private for safer development. (#1155)
 
 ## [0.11.1]
 - Support Scala 3.0.0-RC2